Separate import and validate tests
