+++ tests/data/3.0.1/invalid/package_sbom_missing_creationinfo.json
@@ -1,0 +1,97 @@
+{
+    "@context": "https://spdx.org/rdf/3.0.1/spdx-context.jsonld",
+    "@graph": [
+        {
+            "type": "CreationInfo",
+            "@id": "_:creationinfo",
+            "createdBy": [
+                "http://spdx.example.com/Agent/JoshuaWatt"
+            ],
+            "specVersion": "3.0.1",
+            "created": "2024-03-06T00:00:00Z"
+        },
+        {
+            "type": "Person",
+            "spdxId": "http://spdx.example.com/Agent/JoshuaWatt",
+            "name": "Joshua Watt",
+            "externalIdentifier": [
+                {
+                    "type": "ExternalIdentifier",
+                    "externalIdentifierType": "email",
+                    "identifier": "JPEWhacker@gmail.com"
+                }
+            ]
+        },
+        {
+            "type": "SpdxDocument",
+            "spdxId": "http://spdx.example.com/Document1",
+            "creationInfo": "_:creationinfo",
+            "rootElement": [
+                "http://spdx.example.com/BOM1"
+            ],
+            "element": [
+                "http://spdx.example.com/BOM1",
+                "http://spdx.example.com/Agent/JoshuaWatt",
+                "http://spdx.example.com/Package1/myprogram",
+                "http://spdx.example.com/Relationship/1"
+            ],
+            "profileConformance": [
+                "core",
+                "software"
+            ]
+        },
+        {
+            "type": "software_Sbom",
+            "spdxId": "http://spdx.example.com/BOM1",
+            "creationInfo": "_:creationinfo",
+            "rootElement": [
+                "http://spdx.example.com/Package1"
+            ],
+            "element": [
+                "http://spdx.example.com/Package1/myprogram",
+                "http://spdx.example.com/Package1"
+            ],
+            "software_sbomType": [
+                "build"
+            ]
+        },
+        {
+            "type": "software_Package",
+            "spdxId": "http://spdx.example.com/Package1",
+            "creationInfo": "_:creationinfo",
+            "name": "my-package",
+            "software_packageVersion": "1.0.0",
+            "software_downloadLocation": "http://dl.example.com/my-package_1.0.0.tar",
+            "builtTime": "2024-03-06T00:00:00Z",
+            "originatedBy": [
+                "http://spdx.example.com/Agent/JoshuaWatt"
+            ]
+        },
+        {
+            "type": "software_File",
+            "spdxId": "http://spdx.example.com/Package1/myprogram",
+            "creationInfo": "_:creationinfo",
+            "name": "myprogram",
+            "software_primaryPurpose": "executable",
+            "software_additionalPurpose": [
+                "application"
+            ],
+            "software_copyrightText": "Copyright 2024, Joshua Watt",
+            "builtTime": "2024-03-06T00:00:00Z",
+            "originatedBy": [
+                "http://spdx.example.com/Agent/JoshuaWatt"
+            ]
+        },
+        {
+            "type": "Relationship",
+            "spdxId": "http://spdx.example.com/Relationship/1",
+            "creationInfo": "_:creationinfo",
+            "from": "http://spdx.example.com/Package1",
+            "relationshipType": "contains",
+            "to": [
+                "http://spdx.example.com/Package1/myprogram"
+            ],
+            "completeness": "complete"
+        }
+    ]
+}
